+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.0] - 2024-12-01
+
+### Added
+- Home purchase planning feature
+- Child education planning feature
+- Life insurance needs calculation
+- Financial data visualization in comparison table format
+- User data saving and loading
+
+### Changed
+- Enhanced financial analysis by adding more factors
+
 ## [1.1.0] - 2024-09-06
 
 ### Added
@@ -23,3 +35,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Initial release of Finance Planner
 - Basic financial planning features
+
+## [Unreleased]
+
+### Added
+- Support for Indonesian Rupiah (IDR) in currency conversion feature